multi_model_rag_api/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
openai>=1.0.0
langchain>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
pydantic>=2.0.0
python-multipart>=0.0.5
Pillow>=9.0.0
unstructured[all-docs]>=0.10.0
nest-asyncio>=1.5.0
requests>=2.31.0 